HACK: i2c: s3c2410: stop all pending transfers on shutdown
If the i2c bus is busy when reboot() is called, it may never
come out of the busy state. Usually this is not an issue - since
the hardware will be reset before another kernel comes along - but
this causes the i2c bus to be unusable when using kexec.

Note: the kernel must be in a state where it will not attempt any
accesses over the i2c bus before shutdown. For instance, cpufreq
attempting to access an i2c regulator will not work, and may stop
shutdown from occuring at all.