ASoC: samsung: add sound support to midas boards
This patch adds support for voice and BT calls, along
with standard audio output via the speaker, earpiece, headphone jack,
HDMI, and any accessories compatible with midas boards. This patch
also supports headphone/headset detection and headsets with inline
buttons.

Note that on i9300 and i9305 additional changes
are required to co-operate with the MAX77693 MUIC in order
to use accessory audio output.