leds: an30259a: fix sloping brightness
The DUTYMAX/DUTYMID/DUTYMIN fields represent PWM duty. Previously
I assumed that the high level used by PWM was actually the maximum
current output, but it's actually the current level indicated in the
LEDnCC register - so we don't need to set DUTY*, we can just set
DUTYMAX/MID to "on all the time" to achieve the desired effect.