extcon: max77693: improve interrupt handling logic
When a cable is attached/detached, several interrupts are produced.
In particular, when an ADC cable (like a UART cable) is attached,
ADC and charger interrupts are sent. Currently, the last interrupt to
arrive before the bottom-half handler is executed is the only interrupt
that's actually handled: so it is possible that the ADC interrupt is
never handled, and the extcon state is never updated to reflect the
UART cable's connection.

Fix this by moving the "IRQ type" (ADC or charger) logic into the
interrupt handler, and merely storing the fact that there are pending
ADC/charger interrupts in the driver data. The "bottom-half" handler
now detects cables based on the stored ADC/charger interrupt pending
bits.